mesa: fix float->int mapping for glGetIntegerv(GL_DEPTH_CLEAR_VALUE)
(cherry picked from commit 8124faf89d638285d8e9aa93adc3ca7f4ee729f3)

Conflicts:

	src/mesa/main/get.c

+++ src/mesa/main/get.c
@@ -2137,7 +2137,7 @@ _mesa_GetFloatv( GLenum pname, GLfloat *params )
          params[0] = (GLfloat)(ctx->DrawBuffer->Visual.depthBits);
          break;
       case GL_DEPTH_CLEAR_VALUE:
-         params[0] = (GLfloat)(ctx->Depth.Clear);
+         params[0] = ctx->Depth.Clear;
          break;
       case GL_DEPTH_FUNC:
          params[0] = ENUM_TO_FLOAT(ctx->Depth.Func);
